+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.4.0beta] - 2016-10-xx
 ### Added
 - `bower.json` Optional
+- Set boilerplate version on `gulpfile.js`
 
 ### Changed
 - Moved bower to npm